Refactored nutrition results and its totals
