+++ src/components/Nutrition/nutrition-results.jsx
@@ -3,6 +3,7 @@ import {clearAuth} from '../../actions/auth';
 import {clearAuthToken} from '../../local-storage';
 import '../index.scss';
 import NutritionSearchPage from './nutrition-search-page';
+import NutritionResultsTotals from './nutrition-results-totals';
 import ExerciseSearchPage from './exercise-search-page';
 import {connect} from 'react-redux';
 import {addProtectedData} from '../../actions/protected-data';
@@ -46,50 +47,6 @@ export class NutritionResults extends Component {
     e.preventDefault();
     this.props.dispatch(addProtectedData(nutritionObject, username, date));
     window.alert('You have just saved your nutrition!');
-  }
-
-  renderTotals() {
-    let nutrition_results_array = this.props.nutritionResults;
-    let nutritionTotals = {
-      'food_name': '',
-      'nf_calories': 0,
-      'nf_total_fat': 0,
-      'nf_total_carbohydrate': 0,
-      'nf_protein': 0,
-      'nf_sugars': 0,
-      'nf_sodium': 0
-    };
-    let keys = Object.keys(nutritionTotals);
-    let add_count_of_j;
-
-    for (let i = 0; i < nutrition_results_array.length; i++) {
-      for (let j = 0; j < keys.length; j++) {
-        let foodName = nutrition_results_array[i].food_name.toLowerCase().split(' ').map(s => s.charAt(0).toUpperCase() + s.substring(1)).join(' ');
-        typeof nutritionTotals[keys[j]] === 'string' ? 
-        nutritionTotals[keys[j]] = nutritionTotals[keys[j]] += foodName + (nutrition_results_array[i + 1] ? ', ' : '') : 
-        nutritionTotals[keys[j]] = Math.floor(nutritionTotals[keys[j]] += nutrition_results_array[i][keys[j]]);
-        j === keys.length - 1 ? add_count_of_j = j : console.log('j error');
-      }     
-    }
-
-    if (add_count_of_j) {
-      return (
-        <div className= {!this.props.loading ? "nutrition-results-container fadeIn" : "nutrition-results-container"}>
-          <form onSubmit = {(e) => this.onAdd(e, nutritionTotals, this.props.username, Date.now)}>
-          <h3>Nutrition Totals</h3>
-            <ul className="nutrition-list">
-              <li>{nutritionTotals.nf_calories} calories</li>
-              <li>{nutritionTotals.nf_total_fat} grams of fat</li>
-              <li>{nutritionTotals.nf_total_carbohydrate} grams of carbs</li>
-              <li>{nutritionTotals.nf_protein} grams of protein</li>
-              <li>{nutritionTotals.nf_sugars} grams of sugar</li>
-              <li>{nutritionTotals.nf_sodium} mg of sodium</li>
-            </ul>
-            <button type="submit" className="save-btn">Save Nutrition</button>
-          </form>
-        </div>
-      );
-    }
   }
 
   render() {
@@ -149,7 +106,7 @@ export class NutritionResults extends Component {
         <ul className = {!this.props.loading ? "nutrition-results fadeIn" : "nutrition-results"}>
           {nutrition_result}
         </ul>
-        {this.renderTotals()}
+        <NutritionResultsTotals/>
         {loading}
         {error}
       </section>
